Basii Full Component/Scenario_Generation/requirements_training.txt
openai>=2.14.0
chromadb>=0.4.0
python-dotenv>=1.0.0
flask>=2.3.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0